+++ crates/app/src/tools/feishu.rs
@@ -1,19 +1,13 @@
-#[cfg(test)]
-use std::collections::BTreeMap;
 #[cfg(feature = "tool-file")]
 use std::fs;
-use std::future::Future;
-use std::path::Path;
 use std::sync::{Mutex, OnceLock};
 
 use loong_contracts::{ToolCoreOutcome, ToolCoreRequest};
 use serde::Deserialize;
-use serde::de::DeserializeOwned;
 use serde_json::{Value, json};
 
 use crate::CliResult;
 use crate::channel::ChannelOutboundTarget;
-use crate::channel::feishu::api::resources::bitable;
 use crate::channel::feishu::api::resources::calendar::{
     self, FeishuCalendarFreebusyQuery, FeishuCalendarListQuery,
 };
@@ -39,6 +33,45 @@ const FEISHU_MESSAGE_RESOURCE_ACCEPTED_SCOPES: &[&str] = &[
 const FEISHU_DOC_READ_ACCEPTED_SCOPES: &[&str] = &["docx:document:readonly", "docx:document"];
 const FEISHU_DOC_WRITE_REQUIRED_SCOPES: &[&str] = &["docx:document"];
 const FEISHU_CARD_UPDATE_CALLBACK_TOKEN_USE_LIMIT: usize = 2;
+
+mod bitable_execution;
+mod metadata;
+mod support;
+
+use bitable_execution::*;
+use support::*;
+
+#[cfg(test)]
+pub(super) fn feishu_tool_alias_pairs() -> &'static [(&'static str, &'static str)] {
+    metadata::feishu_tool_alias_pairs()
+}
+
+pub(super) fn canonical_feishu_tool_name(raw: &str) -> Option<&'static str> {
+    metadata::canonical_feishu_tool_name(raw)
+}
+
+pub(super) fn is_known_feishu_tool_name(raw: &str) -> bool {
+    metadata::is_known_feishu_tool_name(raw)
+}
+
+#[cfg(test)]
+pub(super) fn feishu_tool_registry_entries() -> Vec<super::ToolRegistryEntry> {
+    metadata::feishu_tool_registry_entries()
+}
+
+#[cfg(test)]
+pub(super) fn feishu_provider_tool_definitions() -> Vec<Value> {
+    metadata::feishu_provider_tool_definitions()
+}
+
+pub(super) fn feishu_provider_tool_definition(tool_name: &str) -> Option<Value> {
+    metadata::feishu_provider_tool_definition(tool_name)
+}
+
+#[cfg(test)]
+pub(super) fn feishu_shape_examples() -> std::collections::BTreeMap<&'static str, Value> {
+    metadata::feishu_shape_examples()
+}
 
 #[derive(Debug, Clone)]
 pub(crate) struct DeferredFeishuCardUpdate {
@@ -1103,1583 +1136,6 @@ struct FeishuCalendarFreebusyPayload {
     internal: LoongInternalToolPayload,
 }
 
-#[cfg(test)]
-pub(super) fn feishu_tool_alias_pairs() -> &'static [(&'static str, &'static str)] {
-    FEISHU_TOOL_ALIAS_PAIRS
-}
-
-pub(super) fn canonical_feishu_tool_name(raw: &str) -> Option<&'static str> {
-    match raw {
-        "feishu.whoami" | "feishu_whoami" => Some("feishu.whoami"),
-        "feishu.bitable.app.create" | "feishu_bitable_app_create" => {
-            Some("feishu.bitable.app.create")
-        }
-        "feishu.bitable.app.get" | "feishu_bitable_app_get" => Some("feishu.bitable.app.get"),
-        "feishu.bitable.app.list" | "feishu_bitable_app_list" => Some("feishu.bitable.app.list"),
-        "feishu.bitable.app.patch" | "feishu_bitable_app_patch" => Some("feishu.bitable.app.patch"),
-        "feishu.bitable.app.copy" | "feishu_bitable_app_copy" => Some("feishu.bitable.app.copy"),
-        "feishu.bitable.list" | "feishu_bitable_list" => Some("feishu.bitable.list"),
-        "feishu.bitable.table.create" | "feishu_bitable_table_create" => {
-            Some("feishu.bitable.table.create")
-        }
-        "feishu.bitable.table.patch" | "feishu_bitable_table_patch" => {
-            Some("feishu.bitable.table.patch")
-        }
-        "feishu.bitable.table.batch_create" | "feishu_bitable_table_batch_create" => {
-            Some("feishu.bitable.table.batch_create")
-        }
-        "feishu.bitable.record.create" | "feishu_bitable_record_create" => {
-            Some("feishu.bitable.record.create")
-        }
-        "feishu.bitable.record.update" | "feishu_bitable_record_update" => {
-            Some("feishu.bitable.record.update")
-        }
-        "feishu.bitable.record.delete" | "feishu_bitable_record_delete" => {
-            Some("feishu.bitable.record.delete")
-        }
-        "feishu.bitable.record.batch_create" | "feishu_bitable_record_batch_create" => {
-            Some("feishu.bitable.record.batch_create")
-        }
-        "feishu.bitable.record.batch_update" | "feishu_bitable_record_batch_update" => {
-            Some("feishu.bitable.record.batch_update")
-        }
-        "feishu.bitable.record.batch_delete" | "feishu_bitable_record_batch_delete" => {
-            Some("feishu.bitable.record.batch_delete")
-        }
-        "feishu.bitable.field.create" | "feishu_bitable_field_create" => {
-            Some("feishu.bitable.field.create")
-        }
-        "feishu.bitable.field.list" | "feishu_bitable_field_list" => {
-            Some("feishu.bitable.field.list")
-        }
-        "feishu.bitable.field.update" | "feishu_bitable_field_update" => {
-            Some("feishu.bitable.field.update")
-        }
-        "feishu.bitable.field.delete" | "feishu_bitable_field_delete" => {
-            Some("feishu.bitable.field.delete")
-        }
-        "feishu.bitable.view.create" | "feishu_bitable_view_create" => {
-            Some("feishu.bitable.view.create")
-        }
-        "feishu.bitable.view.get" | "feishu_bitable_view_get" => Some("feishu.bitable.view.get"),
-        "feishu.bitable.view.list" | "feishu_bitable_view_list" => Some("feishu.bitable.view.list"),
-        "feishu.bitable.view.patch" | "feishu_bitable_view_patch" => {
-            Some("feishu.bitable.view.patch")
-        }
-        "feishu.bitable.record.search" | "feishu_bitable_record_search" => {
-            Some("feishu.bitable.record.search")
-        }
-        "feishu.doc.create" | "feishu_doc_create" => Some("feishu.doc.create"),
-        "feishu.doc.append" | "feishu_doc_append" => Some("feishu.doc.append"),
-        "feishu.doc.read" | "feishu_doc_read" => Some("feishu.doc.read"),
-        "feishu.messages.history" | "feishu_messages_history" => Some("feishu.messages.history"),
-        "feishu.messages.get" | "feishu_messages_get" => Some("feishu.messages.get"),
-        #[cfg(feature = "tool-file")]
-        "feishu.messages.resource.get" | "feishu_messages_resource_get" => {
-            Some("feishu.messages.resource.get")
-        }
-        "feishu.messages.search" | "feishu_messages_search" => Some("feishu.messages.search"),
-        "feishu.messages.send" | "feishu_messages_send" => Some("feishu.messages.send"),
-        "feishu.messages.reply" | "feishu_messages_reply" => Some("feishu.messages.reply"),
-        "feishu.card.update" | "feishu_card_update" => Some("feishu.card.update"),
-        "feishu.calendar.list" | "feishu_calendar_list" => Some("feishu.calendar.list"),
-        "feishu.calendar.freebusy" | "feishu_calendar_freebusy" => Some("feishu.calendar.freebusy"),
-        "feishu.calendar.primary.get" | "feishu_calendar_primary_get" => {
-            Some("feishu.calendar.primary.get")
-        }
-        _ => None,
-    }
-}
-
-pub(super) fn is_known_feishu_tool_name(raw: &str) -> bool {
-    canonical_feishu_tool_name(raw).is_some()
-}
-
-#[cfg(test)]
-pub(super) fn feishu_tool_registry_entries() -> Vec<super::ToolRegistryEntry> {
-    let mut entries = Vec::new();
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.app.create",
-        "Create a Feishu Bitable app with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.app.get",
-        "Fetch Feishu Bitable app metadata with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.app.list",
-        "List Feishu Bitable apps through the Drive API with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.app.patch",
-        "Update Feishu Bitable app metadata with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.app.copy",
-        "Copy a Feishu Bitable app with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.list",
-        "List data tables in a Feishu Bitable app with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.table.create",
-        "Create a Feishu Bitable table with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.table.patch",
-        "Rename a Feishu Bitable table with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.table.batch_create",
-        "Batch create Feishu Bitable tables with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.record.create",
-        "Create a record in a Feishu Bitable table with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.record.update",
-        "Update a record in a Feishu Bitable table with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.record.delete",
-        "Delete a record in a Feishu Bitable table with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.record.batch_create",
-        "Batch create records in a Feishu Bitable table with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.record.batch_update",
-        "Batch update records in a Feishu Bitable table with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.record.batch_delete",
-        "Batch delete records in a Feishu Bitable table with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.field.create",
-        "Create a field in a Feishu Bitable table with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.field.list",
-        "List fields in a Feishu Bitable table with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.field.update",
-        "Update a field in a Feishu Bitable table with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.field.delete",
-        "Delete a field in a Feishu Bitable table with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.view.create",
-        "Create a view in a Feishu Bitable table with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.view.get",
-        "Fetch a view in a Feishu Bitable table with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.view.list",
-        "List views in a Feishu Bitable table with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.view.patch",
-        "Patch a view in a Feishu Bitable table with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.bitable.record.search",
-        "Search or list records in a Feishu Bitable table with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.calendar.freebusy",
-        "Query Feishu calendar free/busy for the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.doc.create",
-        "Create a Feishu docx document and optionally insert initial markdown or html content with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.doc.append",
-        "Append markdown or html content to an existing Feishu docx document with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.doc.read",
-        "Read Feishu Doc raw content with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.messages.get",
-        "Read one Feishu message detail using a tenant token resolved from the selected account grant",
-    );
-    #[cfg(feature = "tool-file")]
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.messages.resource.get",
-        "Download one Feishu message image or file resource under the configured file root, with safe ingress defaults when the current Feishu turn carries exactly one resource reference",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.messages.history",
-        "List Feishu message history using a tenant token resolved from the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.messages.search",
-        "Search Feishu messages with the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.messages.send",
-        "Send a Feishu text, post, image, file, or markdown card message with a tenant token resolved from the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.messages.reply",
-        "Reply to a Feishu message with text, post, image, file, or a markdown card using a tenant token resolved from the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.card.update",
-        "Update a Feishu interactive card through the delayed callback API, using the current callback token when available",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.calendar.list",
-        "List Feishu calendars or primary calendars for the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.calendar.primary.get",
-        "Fetch the Feishu primary calendar entry for the selected account grant",
-    );
-    push_feishu_registry_entry(
-        &mut entries,
-        "feishu.whoami",
-        "Inspect the active Feishu grant principal and profile",
-    );
-    entries.sort_by(|left, right| left.name.cmp(&right.name));
-    entries
-}
-
-pub(super) fn feishu_provider_tool_definitions() -> Vec<Value> {
-    let mut tools = Vec::new();
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_app_create",
-        "Create a Feishu Bitable app with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string", "description": "Optional Feishu configured account id to route through." },
-                "open_id": { "type": "string", "description": "Optional explicit Feishu user open_id grant selector." },
-                "name": { "type": "string", "description": "Bitable app name." },
-                "folder_token": { "type": "string", "description": "Optional Drive folder token." }
-            },
-            "required": ["name"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_app_get",
-        "Fetch Feishu Bitable app metadata with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string", "description": "Optional Feishu configured account id to route through." },
-                "open_id": { "type": "string", "description": "Optional explicit Feishu user open_id grant selector." },
-                "app_token": { "type": "string", "description": "Feishu Bitable app token." }
-            },
-            "required": ["app_token"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_app_list",
-        "List Feishu Bitable apps through the Drive API with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string", "description": "Optional Feishu configured account id to route through." },
-                "open_id": { "type": "string", "description": "Optional explicit Feishu user open_id grant selector." },
-                "folder_token": { "type": "string", "description": "Optional Drive folder token." },
-                "page_size": { "type": "integer", "minimum": 1, "maximum": 200 },
-                "page_token": { "type": "string" }
-            },
-            "required": [],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_app_patch",
-        "Update Feishu Bitable app metadata with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string", "description": "Optional Feishu configured account id to route through." },
-                "open_id": { "type": "string", "description": "Optional explicit Feishu user open_id grant selector." },
-                "app_token": { "type": "string", "description": "Feishu Bitable app token." },
-                "name": { "type": "string", "description": "Optional new app name." },
-                "is_advanced": { "type": "boolean", "description": "Optional advanced permission toggle." }
-            },
-            "required": ["app_token"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_app_copy",
-        "Copy a Feishu Bitable app with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string", "description": "Optional Feishu configured account id to route through." },
-                "open_id": { "type": "string", "description": "Optional explicit Feishu user open_id grant selector." },
-                "app_token": { "type": "string", "description": "Source Bitable app token." },
-                "name": { "type": "string", "description": "Copied app name." },
-                "folder_token": { "type": "string", "description": "Optional target Drive folder token." }
-            },
-            "required": ["app_token", "name"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_list",
-        "List data tables in a Feishu Bitable app with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": {
-                    "type": "string",
-                    "description": "Optional Feishu configured account id to route through."
-                },
-                "open_id": {
-                    "type": "string",
-                    "description": "Optional explicit Feishu user open_id grant selector."
-                },
-                "app_token": {
-                    "type": "string",
-                    "description": "Feishu Bitable app token."
-                },
-                "page_size": {
-                    "type": "integer",
-                    "minimum": 1,
-                    "maximum": 100
-                },
-                "page_token": {
-                    "type": "string"
-                }
-            },
-            "required": ["app_token"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_table_create",
-        "Create a Feishu Bitable table with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string", "description": "Optional Feishu configured account id to route through." },
-                "open_id": { "type": "string", "description": "Optional explicit Feishu user open_id grant selector." },
-                "app_token": { "type": "string", "description": "Feishu Bitable app token." },
-                "name": { "type": "string", "description": "Bitable table name." },
-                "default_view_name": { "type": "string", "description": "Optional default view name." },
-                "fields": { "type": "array", "items": { "type": "object" }, "description": "Optional table field definitions." }
-            },
-            "required": ["app_token", "name"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_table_patch",
-        "Rename a Feishu Bitable table with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string", "description": "Optional Feishu configured account id to route through." },
-                "open_id": { "type": "string", "description": "Optional explicit Feishu user open_id grant selector." },
-                "app_token": { "type": "string", "description": "Feishu Bitable app token." },
-                "table_id": { "type": "string", "description": "Feishu Bitable table id." },
-                "name": { "type": "string", "description": "New table name." }
-            },
-            "required": ["app_token", "table_id", "name"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_table_batch_create",
-        "Batch create Feishu Bitable tables with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string", "description": "Optional Feishu configured account id to route through." },
-                "open_id": { "type": "string", "description": "Optional explicit Feishu user open_id grant selector." },
-                "app_token": { "type": "string", "description": "Feishu Bitable app token." },
-                "tables": { "type": "array", "items": { "type": "object" }, "description": "Tables to create; only `name` is sent upstream." }
-            },
-            "required": ["app_token", "tables"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_record_create",
-        "Create a record in a Feishu Bitable table with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": {
-                    "type": "string",
-                    "description": "Optional Feishu configured account id to route through."
-                },
-                "open_id": {
-                    "type": "string",
-                    "description": "Optional explicit Feishu user open_id grant selector."
-                },
-                "app_token": {
-                    "type": "string",
-                    "description": "Feishu Bitable app token."
-                },
-                "table_id": {
-                    "type": "string",
-                    "description": "Feishu Bitable table id."
-                },
-                "fields": {
-                    "type": "object",
-                    "description": "Record field values keyed by field name."
-                }
-            },
-            "required": ["app_token", "table_id", "fields"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_record_update",
-        "Update a record in a Feishu Bitable table with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string", "description": "Optional Feishu configured account id to route through." },
-                "open_id": { "type": "string", "description": "Optional explicit Feishu user open_id grant selector." },
-                "app_token": { "type": "string", "description": "Feishu Bitable app token." },
-                "table_id": { "type": "string", "description": "Feishu Bitable table id." },
-                "record_id": { "type": "string", "description": "Feishu Bitable record id." },
-                "fields": { "type": "object", "description": "Record field values keyed by field name." }
-            },
-            "required": ["app_token", "table_id", "record_id", "fields"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_record_delete",
-        "Delete a record in a Feishu Bitable table with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string", "description": "Optional Feishu configured account id to route through." },
-                "open_id": { "type": "string", "description": "Optional explicit Feishu user open_id grant selector." },
-                "app_token": { "type": "string", "description": "Feishu Bitable app token." },
-                "table_id": { "type": "string", "description": "Feishu Bitable table id." },
-                "record_id": { "type": "string", "description": "Feishu Bitable record id." }
-            },
-            "required": ["app_token", "table_id", "record_id"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_record_batch_create",
-        "Batch create records in a Feishu Bitable table with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string" },
-                "open_id": { "type": "string" },
-                "app_token": { "type": "string" },
-                "table_id": { "type": "string" },
-                "records": { "type": "array", "items": { "type": "object" } }
-            },
-            "required": ["app_token", "table_id", "records"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_record_batch_update",
-        "Batch update records in a Feishu Bitable table with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string" },
-                "open_id": { "type": "string" },
-                "app_token": { "type": "string" },
-                "table_id": { "type": "string" },
-                "records": { "type": "array", "items": { "type": "object" } }
-            },
-            "required": ["app_token", "table_id", "records"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_record_batch_delete",
-        "Batch delete records in a Feishu Bitable table with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string" },
-                "open_id": { "type": "string" },
-                "app_token": { "type": "string" },
-                "table_id": { "type": "string" },
-                "records": { "type": "array", "items": { "type": "string" } }
-            },
-            "required": ["app_token", "table_id", "records"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_field_create",
-        "Create a field in a Feishu Bitable table with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string" },
-                "open_id": { "type": "string" },
-                "app_token": { "type": "string" },
-                "table_id": { "type": "string" },
-                "field_name": { "type": "string" },
-                "type": { "type": "integer" },
-                "property": {}
-            },
-            "required": ["app_token", "table_id", "field_name", "type"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_field_list",
-        "List fields in a Feishu Bitable table with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string" },
-                "open_id": { "type": "string" },
-                "app_token": { "type": "string" },
-                "table_id": { "type": "string" },
-                "view_id": { "type": "string" },
-                "page_size": { "type": "integer" },
-                "page_token": { "type": "string" }
-            },
-            "required": ["app_token", "table_id"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_field_update",
-        "Update a field in a Feishu Bitable table with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string" },
-                "open_id": { "type": "string" },
-                "app_token": { "type": "string" },
-                "table_id": { "type": "string" },
-                "field_id": { "type": "string" },
-                "field_name": { "type": "string" },
-                "type": { "type": "integer" },
-                "property": {}
-            },
-            "required": ["app_token", "table_id", "field_id", "field_name", "type"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_field_delete",
-        "Delete a field in a Feishu Bitable table with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string" },
-                "open_id": { "type": "string" },
-                "app_token": { "type": "string" },
-                "table_id": { "type": "string" },
-                "field_id": { "type": "string" }
-            },
-            "required": ["app_token", "table_id", "field_id"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_view_create",
-        "Create a view in a Feishu Bitable table with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string" },
-                "open_id": { "type": "string" },
-                "app_token": { "type": "string" },
-                "table_id": { "type": "string" },
-                "view_name": { "type": "string" },
-                "view_type": { "type": "string" }
-            },
-            "required": ["app_token", "table_id", "view_name"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_view_get",
-        "Fetch a view in a Feishu Bitable table with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string" },
-                "open_id": { "type": "string" },
-                "app_token": { "type": "string" },
-                "table_id": { "type": "string" },
-                "view_id": { "type": "string" }
-            },
-            "required": ["app_token", "table_id", "view_id"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_view_list",
-        "List views in a Feishu Bitable table with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string" },
-                "open_id": { "type": "string" },
-                "app_token": { "type": "string" },
-                "table_id": { "type": "string" },
-                "page_size": { "type": "integer" },
-                "page_token": { "type": "string" }
-            },
-            "required": ["app_token", "table_id"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_view_patch",
-        "Patch a view in a Feishu Bitable table with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": { "type": "string" },
-                "open_id": { "type": "string" },
-                "app_token": { "type": "string" },
-                "table_id": { "type": "string" },
-                "view_id": { "type": "string" },
-                "view_name": { "type": "string" }
-            },
-            "required": ["app_token", "table_id", "view_id", "view_name"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_bitable_record_search",
-        "Search or list records in a Feishu Bitable table with the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": {
-                    "type": "string",
-                    "description": "Optional Feishu configured account id to route through."
-                },
-                "open_id": {
-                    "type": "string",
-                    "description": "Optional explicit Feishu user open_id grant selector."
-                },
-                "app_token": {
-                    "type": "string",
-                    "description": "Feishu Bitable app token."
-                },
-                "table_id": {
-                    "type": "string",
-                    "description": "Feishu Bitable table id."
-                },
-                "view_id": {
-                    "type": "string",
-                    "description": "Optional Bitable view id."
-                },
-                "filter": {
-                    "type": "object",
-                    "description": "Optional Feishu Bitable search filter object."
-                },
-                "sort": {
-                    "type": "array",
-                    "items": {
-                        "type": "object",
-                        "properties": {
-                            "field_name": {
-                                "type": "string"
-                            },
-                            "desc": {
-                                "type": "boolean"
-                            }
-                        },
-                        "required": ["field_name", "desc"],
-                        "additionalProperties": false
-                    },
-                    "description": "Optional Feishu Bitable sort rules."
-                },
-                "field_names": {
-                    "type": "array",
-                    "items": {
-                        "type": "string"
-                    },
-                    "description": "Optional subset of field names to return."
-                },
-                "automatic_fields": {
-                    "type": "boolean",
-                    "description": "Whether to return automatic fields such as created_time and last_modified_time."
-                },
-                "page_size": {
-                    "type": "integer",
-                    "minimum": 1,
-                    "maximum": 500
-                },
-                "page_token": {
-                    "type": "string"
-                }
-            },
-            "required": ["app_token", "table_id"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_calendar_primary_get",
-        "Fetch the Feishu primary calendar entry for the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": {
-                    "type": "string",
-                    "description": "Optional Feishu configured account id to route through."
-                },
-                "open_id": {
-                    "type": "string",
-                    "description": "Optional explicit Feishu user open_id grant selector."
-                },
-                "user_id_type": {
-                    "type": "string",
-                    "description": "Optional Feishu user id type for the response. Defaults to `open_id`."
-                }
-            },
-            "required": [],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_calendar_freebusy",
-        "Query Feishu calendar free/busy for the selected account grant or an explicit user/room.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": {
-                    "type": "string",
-                    "description": "Optional Feishu configured account id to route through."
-                },
-                "open_id": {
-                    "type": "string",
-                    "description": "Optional explicit Feishu user open_id grant selector."
-                },
-                "user_id_type": {
-                    "type": "string",
-                    "description": "Optional Feishu calendar user id type. Defaults to `open_id` when user_id is inferred from the selected grant."
-                },
-                "time_min": {
-                    "type": "string",
-                    "description": "Inclusive time window start, typically RFC3339."
-                },
-                "time_max": {
-                    "type": "string",
-                    "description": "Exclusive time window end, typically RFC3339."
-                },
-                "user_id": {
-                    "type": "string",
-                    "description": "Optional explicit Feishu user id. Defaults to the selected grant open_id when room_id is omitted."
-                },
-                "room_id": {
-                    "type": "string",
-                    "description": "Optional meeting room id to query instead of a user calendar."
-                },
-                "include_external_calendar": {
-                    "type": "boolean",
-                    "description": "Whether to include external calendars."
-                },
-                "only_busy": {
-                    "type": "boolean",
-                    "description": "Whether to return only busy slots."
-                },
-                "need_rsvp_status": {
-                    "type": "boolean",
-                    "description": "Whether to include RSVP status in each slot."
-                }
-            },
-            "required": ["time_min", "time_max"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_card_update",
-        "Update a Feishu interactive card after a card callback. Pass markdown for a standard markdown card or card for full Feishu card JSON. When called from a Feishu callback turn, Loong can infer account_id, callback_token, and a default exclusive open_ids target from internal callback context. Set shared=true for shared-card updates so callback operator defaults are suppressed. Callback tokens expire after 30 minutes and can be used at most twice.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": {
-                    "type": "string",
-                    "description": "Optional Feishu configured account id to route through. Defaults from the current Feishu ingress when available."
-                },
-                "callback_token": {
-                    "type": "string",
-                    "description": "Optional callback token for delayed card updates. Usually inferred from the current Feishu card callback turn."
-                },
-                "card": {
-                    "type": "object",
-                    "description": "Optional full Feishu card JSON object to apply to the existing card. Mutually exclusive with `markdown`."
-                },
-                "markdown": {
-                    "type": "string",
-                    "description": "Optional markdown text to wrap in a standard markdown card. Mutually exclusive with `card`."
-                },
-                "shared": {
-                    "type": "boolean",
-                    "description": "Set true for shared-card updates. Shared-card updates must not send non-empty open_ids, and in callback turns this suppresses the default operator open_id target."
-                },
-                "open_ids": {
-                    "type": "array",
-                    "items": {
-                        "type": "string"
-                    },
-                    "description": "Optional explicit open_id targets for non-shared cards. For shared cards, either omit open_ids or set shared=true. When omitted in a callback turn without shared=true, Loong can default to the callback operator open_id."
-                }
-            },
-            "required": [],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_calendar_list",
-        "List Feishu calendars or primary calendars for the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": {
-                    "type": "string",
-                    "description": "Optional Feishu configured account id to route through."
-                },
-                "open_id": {
-                    "type": "string",
-                    "description": "Optional explicit Feishu user open_id grant selector."
-                },
-                "primary": {
-                    "type": "boolean",
-                    "description": "When true, list primary calendars for the selected user."
-                },
-                "user_id_type": {
-                    "type": "string",
-                    "description": "Optional Feishu user id type for primary calendar lookup."
-                },
-                "page_size": {
-                    "type": "integer",
-                    "minimum": 1,
-                    "maximum": 100
-                },
-                "page_token": {
-                    "type": "string"
-                },
-                "sync_token": {
-                    "type": "string"
-                }
-            },
-            "required": [],
-            "additionalProperties": false
-        }),
-    );
-    let mut doc_create_parameters = json!({
-        "type": "object",
-        "properties": {
-            "account_id": {
-                "type": "string",
-                "description": "Optional Feishu configured account id to route through."
-            },
-            "open_id": {
-                "type": "string",
-                "description": "Optional explicit Feishu user open_id grant selector."
-            },
-            "title": {
-                "type": "string",
-                "description": "Optional plain-text Feishu document title."
-            },
-            "folder_token": {
-                "type": "string",
-                "description": "Optional folder token where the new document should be created."
-            },
-            "content": {
-                "type": "string",
-                "description": "Optional initial content to convert and insert into the new document. Mutually exclusive with `content_path`."
-            },
-            "content_type": {
-                "type": "string",
-                "enum": ["markdown", "html"],
-                "description": "Optional content format for `content` or `content_path`. Defaults to the file extension for `content_path` (`.md`/`.markdown` => markdown, `.html`/`.htm` => html) and otherwise `markdown`."
-            }
-        },
-        "required": [],
-        "additionalProperties": false
-    });
-    #[cfg(feature = "tool-file")]
-    if let Some(properties) = doc_create_parameters
-        .get_mut("properties")
-        .and_then(Value::as_object_mut)
-    {
-        properties.insert(
-            "content_path".to_owned(),
-            json!({
-                "type": "string",
-                "description": "Optional relative or rooted local UTF-8 text file path resolved under the configured tool file root and inserted into the new document. Mutually exclusive with `content`."
-            }),
-        );
-    }
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_doc_create",
-        "Create a Feishu document with the selected account grant and optionally insert initial markdown or html content into the new doc.",
-        doc_create_parameters,
-    );
-    let mut doc_append_parameters = json!({
-        "type": "object",
-        "properties": {
-            "account_id": {
-                "type": "string",
-                "description": "Optional Feishu configured account id to route through."
-            },
-            "open_id": {
-                "type": "string",
-                "description": "Optional explicit Feishu user open_id grant selector."
-            },
-            "url": {
-                "type": "string",
-                "description": "Feishu docx URL or document id of the existing document to append to."
-            },
-            "content": {
-                "type": "string",
-                "description": "Markdown or html content to convert and append to the document. Mutually exclusive with `content_path`."
-            },
-            "content_type": {
-                "type": "string",
-                "enum": ["markdown", "html"],
-                "description": "Optional content format for `content` or `content_path`. Defaults to the file extension for `content_path` (`.md`/`.markdown` => markdown, `.html`/`.htm` => html) and otherwise `markdown`."
-            }
-        },
-        "required": ["url", "content"],
-        "additionalProperties": false
-    });
-    #[cfg(feature = "tool-file")]
-    if let Some(parameters) = doc_append_parameters.as_object_mut() {
-        if let Some(properties) = parameters
-            .get_mut("properties")
-            .and_then(Value::as_object_mut)
-        {
-            properties.insert(
-                "content_path".to_owned(),
-                json!({
-                    "type": "string",
-                    "description": "Relative or rooted local UTF-8 text file path resolved under the configured tool file root and appended to the document. Mutually exclusive with `content`."
-                }),
-            );
-        }
-        parameters.insert("required".to_owned(), json!(["url"]));
-        parameters.insert(
-            "anyOf".to_owned(),
-            json!([
-                { "required": ["content"] },
-                { "required": ["content_path"] }
-            ]),
-        );
-    }
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_doc_append",
-        "Append markdown or html content to an existing Feishu document identified by docx url or document id using the selected account grant.",
-        doc_append_parameters,
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_doc_read",
-        "Read a Feishu document by docx url or document id using the selected account grant.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": {
-                    "type": "string",
-                    "description": "Optional Feishu configured account id to route through."
-                },
-                "open_id": {
-                    "type": "string",
-                    "description": "Optional explicit Feishu user open_id grant selector."
-                },
-                "url": {
-                    "type": "string",
-                    "description": "Feishu docx URL or document id."
-                },
-                "lang": {
-                    "type": "integer",
-                    "minimum": 0,
-                    "maximum": 255,
-                    "description": "Optional Feishu language selector."
-                }
-            },
-            "required": ["url"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_messages_get",
-        "Fetch one Feishu message detail using a tenant token resolved from the selected account grant. When called from a Feishu conversation, Loong can infer the account and current message from ingress context.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": {
-                    "type": "string",
-                    "description": "Optional Feishu configured account id to route through."
-                },
-                "open_id": {
-                    "type": "string",
-                    "description": "Optional explicit Feishu user open_id grant selector."
-                },
-                "message_id": {
-                    "type": "string",
-                    "description": "Feishu message id to fetch. Optional when current Feishu ingress already provides the source message id."
-                }
-            },
-            "required": [],
-            "additionalProperties": false
-        }),
-    );
-    #[cfg(feature = "tool-file")]
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_messages_resource_get",
-        "Explicitly download one Feishu message image or file resource using a tenant token resolved from the selected account grant and save it under the configured file root. When called from a Feishu conversation, Loong can infer the source message from ingress context and can infer the resource key or type when the current Feishu ingress carries exactly one Feishu message resource or when either payload.file_key or payload.type uniquely identifies one current ingress resource for the same message, as long as payload.message_id is omitted or matches the current ingress message. If the current Feishu ingress summary exposes resource_inventory, choose one entry and copy its file_key plus payload_type into this tool call when multiple resources are present. Outside the current ingress turn, also pass the source message_id explicitly. This does not perform implicit webhook binary downloads.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": {
-                    "type": "string",
-                    "description": "Optional Feishu configured account id to route through."
-                },
-                "open_id": {
-                    "type": "string",
-                    "description": "Optional explicit Feishu user open_id grant selector."
-                },
-                "message_id": {
-                    "type": "string",
-                    "description": "Feishu message id that owns the resource. Optional when current Feishu ingress already identifies the source message. Outside the current ingress turn, provide this explicitly. If you override it to a different message, current ingress resource defaults no longer apply."
-                },
-                "file_key": {
-                    "type": "string",
-                    "description": "Feishu message resource key paired with the source message id. Optional when the current Feishu ingress carries exactly one Feishu message resource or when payload.type uniquely selects one current ingress resource for the same message, as long as payload.message_id is omitted or matches the current ingress message. If the current Feishu ingress summary includes resource_inventory and multiple resources are present, choose one entry and copy its file_key explicitly."
-                },
-                "type": {
-                    "type": "string",
-                    "enum": ["image", "file", "audio", "media"],
-                    "description": "Feishu message resource type. Use `image` for image resources, preview images from media messages, and image resource keys; use `file`, `audio`, or `media` for binary file resources. `audio` and `media` aliases normalize to the Feishu file transport type. If the current Feishu ingress summary includes resource_inventory, copy the selected entry's payload_type here. Optional when the current Feishu ingress carries exactly one Feishu message resource or when payload.file_key uniquely selects one current ingress resource for the same message, as long as payload.message_id is omitted or matches the current ingress message."
-                },
-                "save_as": {
-                    "type": "string",
-                    "description": "Relative file path to write under the configured file root."
-                }
-            },
-            "required": ["save_as"],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_messages_history",
-        "List Feishu message history using a tenant token resolved from the selected account grant. When called from a Feishu conversation, Loong can infer the current chat or thread container from ingress context.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": {
-                    "type": "string",
-                    "description": "Optional Feishu configured account id to route through."
-                },
-                "open_id": {
-                    "type": "string",
-                    "description": "Optional explicit Feishu user open_id grant selector."
-                },
-                "container_id_type": {
-                    "type": "string",
-                    "description": "Feishu message container id type, for example `chat` or `thread`. Optional when current Feishu conversation ingress can infer the active chat or thread."
-                },
-                "container_id": {
-                    "type": "string",
-                    "description": "Feishu message container id. Optional when current Feishu conversation ingress can infer the active chat or thread id."
-                },
-                "start_time": {
-                    "type": "string"
-                },
-                "end_time": {
-                    "type": "string"
-                },
-                "sort_type": {
-                    "type": "string"
-                },
-                "page_size": {
-                    "type": "integer",
-                    "minimum": 1,
-                    "maximum": 100
-                },
-                "page_token": {
-                    "type": "string"
-                }
-            },
-            "required": [],
-            "additionalProperties": false
-        }),
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_messages_search",
-        "Search Feishu messages using the selected account grant. When called from the current Feishu conversation, Loong can infer the account and default chat scope from ingress context.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": {
-                    "type": "string",
-                    "description": "Optional Feishu configured account id to route through."
-                },
-                "open_id": {
-                    "type": "string",
-                    "description": "Optional explicit Feishu user open_id grant selector."
-                },
-                "user_id_type": {
-                    "type": "string",
-                    "description": "Optional Feishu search user id type."
-                },
-                "page_size": {
-                    "type": "integer",
-                    "minimum": 1,
-                    "maximum": 100
-                },
-                "page_token": {
-                    "type": "string"
-                },
-                "query": {
-                    "type": "string",
-                    "description": "Search query string."
-                },
-                "from_ids": {
-                    "type": "array",
-                    "items": {"type": "string"}
-                },
-                "chat_ids": {
-                    "type": "array",
-                    "items": {"type": "string"},
-                    "description": "Optional Feishu chat ids to scope the search. When omitted inside the current Feishu conversation, Loong can default this to the active conversation."
-                },
-                "message_type": {
-                    "type": "string"
-                },
-                "at_chatter_ids": {
-                    "type": "array",
-                    "items": {"type": "string"}
-                },
-                "from_type": {
-                    "type": "string"
-                },
-                "chat_type": {
-                    "type": "string"
-                },
-                "start_time": {
-                    "type": "string"
-                },
-                "end_time": {
-                    "type": "string"
-                }
-            },
-            "required": ["query"],
-            "additionalProperties": false
-        }),
-    );
-    let mut send_parameters = json!({
-        "type": "object",
-        "properties": {
-            "account_id": {
-                "type": "string",
-                "description": "Optional Feishu configured account id to route through."
-            },
-            "open_id": {
-                "type": "string",
-                "description": "Optional explicit Feishu user open_id grant selector."
-            },
-            "receive_id_type": {
-                "type": "string",
-                "description": "Optional Feishu receive_id_type override. Defaults to the configured account receive_id_type."
-            },
-            "receive_id": {
-                "type": "string",
-                "description": "Feishu receive id to send to. Optional when current Feishu conversation ingress already identifies the active chat."
-            },
-            "text": {
-                "type": "string",
-                "description": "Optional plain-text body to send. Mutually exclusive with `post`, `image_key`, `image_path`, `file_key`, and `file_path`."
-            },
-            "post": {
-                "type": "object",
-                "description": "Optional Feishu post rich-text content JSON object. Mutually exclusive with `text`, `image_key`, `image_path`, `file_key`, and `file_path`; incompatible with `as_card`."
-            },
-            "image_key": {
-                "type": "string",
-                "description": "Optional uploaded Feishu image_key for an image message. Mutually exclusive with `text`, `post`, `image_path`, `file_key`, `file_path`, and `as_card`."
-            },
-            "file_key": {
-                "type": "string",
-                "description": "Optional uploaded Feishu file_key for a file message. Mutually exclusive with `text`, `post`, `image_key`, `image_path`, `file_path`, and `as_card`."
-            },
-            "as_card": {
-                "type": "boolean",
-                "description": "When true, wrap `text` in a markdown interactive card instead of sending a plain text message. Not allowed with `post`, `image_key`, `image_path`, `file_key`, or `file_path`."
-            },
-            "uuid": {
-                "type": "string",
-                "description": "Optional Feishu request UUID used for one-hour message deduplication."
-            }
-        },
-        "required": [],
-        "additionalProperties": false
-    });
-    #[cfg(feature = "tool-file")]
-    if let Some(properties) = send_parameters
-        .get_mut("properties")
-        .and_then(Value::as_object_mut)
-    {
-        properties.insert(
-            "image_path".to_owned(),
-            json!({
-                "type": "string",
-                "description": "Optional relative or rooted local image path resolved under the configured tool file root, uploaded to Feishu before sending. Mutually exclusive with `image_key`, `text`, `post`, `file_key`, `file_path`, and `as_card`."
-            }),
-        );
-        properties.insert(
-            "file_path".to_owned(),
-            json!({
-                "type": "string",
-                "description": "Optional relative or rooted local file path resolved under the configured tool file root, uploaded to Feishu before sending. Mutually exclusive with `file_key`, `text`, `post`, `image_key`, `image_path`, and `as_card`."
-            }),
-        );
-        properties.insert(
-            "file_type".to_owned(),
-            json!({
-                "type": "string",
-                "description": "Optional Feishu upload file_type used only with `file_path`. Defaults to `stream`."
-            }),
-        );
-    }
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_messages_send",
-        "Send a Feishu text, post, image, file, or markdown card message using a tenant token resolved from the selected account grant. When called from the current Feishu conversation, Loong can infer the account and receive_id from ingress context.",
-        send_parameters,
-    );
-    let mut reply_parameters = json!({
-        "type": "object",
-        "properties": {
-            "account_id": {
-                "type": "string",
-                "description": "Optional Feishu configured account id to route through."
-            },
-            "open_id": {
-                "type": "string",
-                "description": "Optional explicit Feishu user open_id grant selector."
-            },
-            "message_id": {
-                "type": "string",
-                "description": "Feishu message id to reply to. Optional when current Feishu ingress already identifies the source Feishu message."
-            },
-            "text": {
-                "type": "string",
-                "description": "Optional plain-text reply body. Mutually exclusive with `post`, `image_key`, `image_path`, `file_key`, and `file_path`."
-            },
-            "post": {
-                "type": "object",
-                "description": "Optional Feishu post rich-text content JSON object. Mutually exclusive with `text`, `image_key`, `image_path`, `file_key`, and `file_path`; incompatible with `as_card`."
-            },
-            "image_key": {
-                "type": "string",
-                "description": "Optional uploaded Feishu image_key for an image reply. Mutually exclusive with `text`, `post`, `image_path`, `file_key`, `file_path`, and `as_card`."
-            },
-            "file_key": {
-                "type": "string",
-                "description": "Optional uploaded Feishu file_key for a file reply. Mutually exclusive with `text`, `post`, `image_key`, `image_path`, `file_path`, and `as_card`."
-            },
-            "as_card": {
-                "type": "boolean",
-                "description": "When true, wrap `text` in a markdown interactive card. Not allowed with `post`, `image_key`, `image_path`, `file_key`, or `file_path`."
-            },
-            "reply_in_thread": {
-                "type": "boolean",
-                "description": "When true, force the reply to be posted in thread form. When omitted, Loong defaults to thread form if internal Feishu ingress metadata indicates the source message is already in a thread/topic."
-            },
-            "uuid": {
-                "type": "string",
-                "description": "Optional Feishu request UUID used for one-hour reply deduplication."
-            }
-        },
-        "required": [],
-        "additionalProperties": false
-    });
-    #[cfg(feature = "tool-file")]
-    if let Some(properties) = reply_parameters
-        .get_mut("properties")
-        .and_then(Value::as_object_mut)
-    {
-        properties.insert(
-            "image_path".to_owned(),
-            json!({
-                "type": "string",
-                "description": "Optional relative or rooted local image path resolved under the configured tool file root, uploaded to Feishu before replying. Mutually exclusive with `image_key`, `text`, `post`, `file_key`, `file_path`, and `as_card`."
-            }),
-        );
-        properties.insert(
-            "file_path".to_owned(),
-            json!({
-                "type": "string",
-                "description": "Optional relative or rooted local file path resolved under the configured tool file root, uploaded to Feishu before replying. Mutually exclusive with `file_key`, `text`, `post`, `image_key`, `image_path`, and `as_card`."
-            }),
-        );
-        properties.insert(
-            "file_type".to_owned(),
-            json!({
-                "type": "string",
-                "description": "Optional Feishu upload file_type used only with `file_path`. Defaults to `stream`."
-            }),
-        );
-    }
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_messages_reply",
-        "Reply to a Feishu message with text, post, image, file, or a markdown card using a tenant token resolved from the selected account grant. When called from a Feishu conversation, Loong can infer the account and source Feishu message from ingress context.",
-        reply_parameters,
-    );
-    push_feishu_provider_tool_definition(
-        &mut tools,
-        "feishu_whoami",
-        "Resolve the currently selected Feishu OAuth grant and fetch the live user profile.",
-        json!({
-            "type": "object",
-            "properties": {
-                "account_id": {
-                    "type": "string",
-                    "description": "Optional Feishu configured account id to route through."
-                },
-                "open_id": {
-                    "type": "string",
-                    "description": "Optional explicit Feishu user open_id grant selector."
-                }
-            },
-            "required": [],
-            "additionalProperties": false
-        }),
-    );
-    tools.sort_by(|left, right| {
-        feishu_provider_tool_function_name(left).cmp(feishu_provider_tool_function_name(right))
-    });
-    tools
-}
-
-pub(super) fn feishu_provider_tool_definition(tool_name: &str) -> Option<Value> {
-    feishu_provider_tool_definitions()
-        .into_iter()
-        .find(|definition| {
-            definition
-                .get("function")
-                .and_then(|value| value.get("name"))
-                .and_then(Value::as_str)
-                .map(super::canonical_tool_name)
-                == Some(tool_name)
-        })
-}
-
-#[cfg(test)]
-pub(super) fn feishu_shape_examples() -> BTreeMap<&'static str, Value> {
-    let mut shapes = BTreeMap::new();
-    shapes.insert(
-        "feishu.bitable.list",
-        json!({
-            "app_token": "bascnDemoAppToken",
-            "page_size": 20
-        }),
-    );
-    shapes.insert(
-        "feishu.bitable.record.create",
-        json!({
-            "app_token": "bascnDemoAppToken",
-            "table_id": "tblDemo",
-            "fields": {
-                "Name": "Release note",
-                "Status": "Draft"
-            }
-        }),
-    );
-    shapes.insert(
-        "feishu.bitable.record.search",
-        json!({
-            "app_token": "bascnDemoAppToken",
-            "table_id": "tblDemo",
-            "page_size": 20
-        }),
-    );
-    shapes.insert(
-        "feishu.doc.create",
-        json!({
-            "title": "Release Plan",
-            "content": "# Release Plan",
-            "content_type": "markdown"
-        }),
-    );
-    shapes.insert(
-        "feishu.doc.read",
-        json!({
-            "url": "https://open.feishu.cn/docx/doxcnDemo"
-        }),
-    );
-    shapes.insert(
-        "feishu.doc.append",
-        json!({
-            "url": "https://open.feishu.cn/docx/doxcnDemo",
-            "content": "Follow-up note"
-        }),
-    );
-    shapes.insert(
-        "feishu.messages.search",
-        json!({
-            "query": "release note",
-            "chat_ids": ["oc_demo_chat"]
-        }),
-    );
-    shapes.insert(
-        "feishu.messages.history",
-        json!({
-            "container_id_type": "chat",
-            "container_id": "oc_demo_chat",
-            "page_size": 20
-        }),
-    );
-    shapes.insert(
-        "feishu.messages.get",
-        json!({
-            "message_id": "om_123"
-        }),
-    );
-    #[cfg(feature = "tool-file")]
-    shapes.insert(
-        "feishu.messages.send",
-        json!({
-            "receive_id": "oc_demo_chat",
-            "image_path": "uploads/demo.png"
-        }),
-    );
-    #[cfg(feature = "tool-file")]
-    shapes.insert(
-        "feishu.messages.reply",
-        json!({
-            "message_id": "om_123",
-            "file_path": "uploads/spec-sheet.pdf",
-            "file_type": "stream"
-        }),
-    );
-    #[cfg(feature = "tool-file")]
-    shapes.insert(
-        "feishu.messages.resource.get",
-        json!({
-            "message_id": "om_123",
-            "file_key": "img_from_resource_inventory",
-            "type": "image",
-            "save_as": "downloads/preview.png"
-        }),
-    );
-    shapes.insert(
-        "feishu.card.update",
-        json!({
-            "shared": true,
-            "markdown": "Approved for everyone"
-        }),
-    );
-    shapes.insert(
-        "feishu.calendar.list",
-        json!({
-            "primary": true
-        }),
-    );
-    shapes.insert(
-        "feishu.calendar.freebusy",
-        json!({
-            "time_min": "2026-03-12T09:00:00+08:00",
-            "time_max": "2026-03-12T18:00:00+08:00"
-        }),
-    );
-    shapes.insert("feishu.whoami", json!({}));
-    shapes
-}
-
 pub(super) fn execute_feishu_tool_with_config(
     request: ToolCoreRequest,
     config: &super::runtime_config::ToolRuntimeConfig,
@@ -2814,1239 +1270,6 @@ fn execute_feishu_whoami_tool_with_config(
                 "user_info": user_info,
                 "grant_scopes": grant.scopes.as_slice(),
             }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_list_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload =
-        parse_payload::<FeishuBitableListPayload>("feishu.bitable.list", request.payload)?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty("feishu.bitable.list", "app_token", &payload.app_token)?;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_required_scopes(&grant, &["base:table:read"], tool_name.as_str())?;
-        let result = bitable::list_bitable_tables(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            payload.page_token.as_deref(),
-            payload.page_size,
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({
-                "tables": result.items,
-                "has_more": result.has_more,
-                "page_token": result.page_token,
-            }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_app_create_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload = parse_payload::<FeishuBitableAppCreatePayload>(
-        "feishu.bitable.app.create",
-        request.payload,
-    )?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let name = require_non_empty("feishu.bitable.app.create", "name", &payload.name)?;
-    let folder_token = payload.folder_token;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["bitable:app"], tool_name.as_str())?;
-        let app = bitable::create_bitable_app(
-            &context.client,
-            &grant.access_token,
-            &name,
-            folder_token.as_deref(),
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({ "app": app }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_app_get_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload =
-        parse_payload::<FeishuBitableAppGetPayload>("feishu.bitable.app.get", request.payload)?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty("feishu.bitable.app.get", "app_token", &payload.app_token)?;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["bitable:app"], tool_name.as_str())?;
-        let app =
-            bitable::get_bitable_app(&context.client, &grant.access_token, &app_token).await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({ "app": app }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_app_list_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload =
-        parse_payload::<FeishuBitableAppListPayload>("feishu.bitable.app.list", request.payload)?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let query = bitable::BitableAppListQuery {
-        folder_token: payload.folder_token,
-        page_token: payload.page_token,
-        page_size: payload.page_size,
-    };
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_required_scopes(&grant, &["drive:drive:readonly"], tool_name.as_str())?;
-        let result =
-            bitable::list_bitable_apps(&context.client, &grant.access_token, &query).await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({
-                "apps": result.apps,
-                "page_token": result.page_token,
-                "has_more": result.has_more,
-            }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_app_patch_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload =
-        parse_payload::<FeishuBitableAppPatchPayload>("feishu.bitable.app.patch", request.payload)?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty("feishu.bitable.app.patch", "app_token", &payload.app_token)?;
-    let name = payload.name;
-    let is_advanced = payload.is_advanced;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["bitable:app"], tool_name.as_str())?;
-        let app = bitable::patch_bitable_app(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            name.as_deref(),
-            is_advanced,
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({ "app": app }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_app_copy_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload =
-        parse_payload::<FeishuBitableAppCopyPayload>("feishu.bitable.app.copy", request.payload)?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty("feishu.bitable.app.copy", "app_token", &payload.app_token)?;
-    let name = require_non_empty("feishu.bitable.app.copy", "name", &payload.name)?;
-    let folder_token = payload.folder_token;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["bitable:app"], tool_name.as_str())?;
-        let app = bitable::copy_bitable_app(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            &name,
-            folder_token.as_deref(),
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({ "app": app }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_record_create_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload = parse_payload::<FeishuBitableRecordCreatePayload>(
-        "feishu.bitable.record.create",
-        request.payload,
-    )?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty(
-        "feishu.bitable.record.create",
-        "app_token",
-        &payload.app_token,
-    )?;
-    let table_id = require_non_empty(
-        "feishu.bitable.record.create",
-        "table_id",
-        &payload.table_id,
-    )?;
-    if !payload.fields.is_object() {
-        return Err(format!(
-            "feishu.bitable.record.create: `fields` must be a JSON object, got {}",
-            payload.fields
-        ));
-    }
-    let fields = payload.fields;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_required_scopes(&grant, &["base:record:create"], tool_name.as_str())?;
-        let record = bitable::create_bitable_record(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            &table_id,
-            fields,
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({
-                "record": record,
-            }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_table_create_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload = parse_payload::<FeishuBitableTableCreatePayload>(
-        "feishu.bitable.table.create",
-        request.payload,
-    )?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty(
-        "feishu.bitable.table.create",
-        "app_token",
-        &payload.app_token,
-    )?;
-    let name = require_non_empty("feishu.bitable.table.create", "name", &payload.name)?;
-    let default_view_name = payload.default_view_name;
-    let fields = payload.fields;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["bitable:app"], tool_name.as_str())?;
-        let result = bitable::create_bitable_table(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            &name,
-            default_view_name.as_deref(),
-            fields,
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({ "result": result }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_table_patch_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload = parse_payload::<FeishuBitableTablePatchPayload>(
-        "feishu.bitable.table.patch",
-        request.payload,
-    )?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty(
-        "feishu.bitable.table.patch",
-        "app_token",
-        &payload.app_token,
-    )?;
-    let table_id = require_non_empty("feishu.bitable.table.patch", "table_id", &payload.table_id)?;
-    let name = require_non_empty("feishu.bitable.table.patch", "name", &payload.name)?;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["bitable:app"], tool_name.as_str())?;
-        let result = bitable::patch_bitable_table(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            &table_id,
-            &name,
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({ "result": result }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_table_batch_create_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload = parse_payload::<FeishuBitableTableBatchCreatePayload>(
-        "feishu.bitable.table.batch_create",
-        request.payload,
-    )?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty(
-        "feishu.bitable.table.batch_create",
-        "app_token",
-        &payload.app_token,
-    )?;
-    let tables = payload.tables;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["bitable:app"], tool_name.as_str())?;
-        let result = bitable::batch_create_bitable_tables(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            tables,
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({ "result": result }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_record_search_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload = parse_payload::<FeishuBitableRecordSearchPayload>(
-        "feishu.bitable.record.search",
-        request.payload,
-    )?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty(
-        "feishu.bitable.record.search",
-        "app_token",
-        &payload.app_token,
-    )?;
-    let table_id = require_non_empty(
-        "feishu.bitable.record.search",
-        "table_id",
-        &payload.table_id,
-    )?;
-    let query = bitable::BitableRecordSearchQuery {
-        page_token: payload.page_token,
-        page_size: payload.page_size,
-        view_id: payload.view_id,
-        filter: payload.filter,
-        sort: payload.sort,
-        field_names: payload.field_names,
-        automatic_fields: payload.automatic_fields,
-    };
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_required_scopes(&grant, &["base:record:retrieve"], tool_name.as_str())?;
-        let result = bitable::search_bitable_records(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            &table_id,
-            &query,
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({
-                "result": result,
-            }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_record_update_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload = parse_payload::<FeishuBitableRecordUpdatePayload>(
-        "feishu.bitable.record.update",
-        request.payload,
-    )?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty(
-        "feishu.bitable.record.update",
-        "app_token",
-        &payload.app_token,
-    )?;
-    let table_id = require_non_empty(
-        "feishu.bitable.record.update",
-        "table_id",
-        &payload.table_id,
-    )?;
-    let record_id = require_non_empty(
-        "feishu.bitable.record.update",
-        "record_id",
-        &payload.record_id,
-    )?;
-    if !payload.fields.is_object() {
-        return Err(format!(
-            "feishu.bitable.record.update: `fields` must be a JSON object, got {}",
-            payload.fields
-        ));
-    }
-    let fields = payload.fields;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["base:record:write"], tool_name.as_str())?;
-        let record = bitable::update_bitable_record(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            &table_id,
-            &record_id,
-            fields,
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({ "record": record }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_record_delete_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload = parse_payload::<FeishuBitableRecordDeletePayload>(
-        "feishu.bitable.record.delete",
-        request.payload,
-    )?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty(
-        "feishu.bitable.record.delete",
-        "app_token",
-        &payload.app_token,
-    )?;
-    let table_id = require_non_empty(
-        "feishu.bitable.record.delete",
-        "table_id",
-        &payload.table_id,
-    )?;
-    let record_id = require_non_empty(
-        "feishu.bitable.record.delete",
-        "record_id",
-        &payload.record_id,
-    )?;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["base:record:write"], tool_name.as_str())?;
-        let result = bitable::delete_bitable_record(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            &table_id,
-            &record_id,
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({
-                "deleted": result.deleted,
-                "record_id": result.record_id,
-            }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_record_batch_create_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload = parse_payload::<FeishuBitableRecordBatchCreatePayload>(
-        "feishu.bitable.record.batch_create",
-        request.payload,
-    )?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty(
-        "feishu.bitable.record.batch_create",
-        "app_token",
-        &payload.app_token,
-    )?;
-    let table_id = require_non_empty(
-        "feishu.bitable.record.batch_create",
-        "table_id",
-        &payload.table_id,
-    )?;
-    let records = payload.records;
-    bitable::ensure_bitable_batch_limit("feishu.bitable.record.batch_create", records.len())?;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["base:record:write"], tool_name.as_str())?;
-        let result = bitable::batch_create_bitable_records(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            &table_id,
-            records,
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({ "result": result }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_record_batch_update_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload = parse_payload::<FeishuBitableRecordBatchUpdatePayload>(
-        "feishu.bitable.record.batch_update",
-        request.payload,
-    )?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty(
-        "feishu.bitable.record.batch_update",
-        "app_token",
-        &payload.app_token,
-    )?;
-    let table_id = require_non_empty(
-        "feishu.bitable.record.batch_update",
-        "table_id",
-        &payload.table_id,
-    )?;
-    let records = payload.records;
-    bitable::ensure_bitable_batch_limit("feishu.bitable.record.batch_update", records.len())?;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["base:record:write"], tool_name.as_str())?;
-        let result = bitable::batch_update_bitable_records(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            &table_id,
-            records,
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({ "result": result }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_record_batch_delete_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload = parse_payload::<FeishuBitableRecordBatchDeletePayload>(
-        "feishu.bitable.record.batch_delete",
-        request.payload,
-    )?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty(
-        "feishu.bitable.record.batch_delete",
-        "app_token",
-        &payload.app_token,
-    )?;
-    let table_id = require_non_empty(
-        "feishu.bitable.record.batch_delete",
-        "table_id",
-        &payload.table_id,
-    )?;
-    let records = payload.records;
-    bitable::ensure_bitable_batch_limit("feishu.bitable.record.batch_delete", records.len())?;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["base:record:write"], tool_name.as_str())?;
-        let result = bitable::batch_delete_bitable_records(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            &table_id,
-            records,
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({ "result": result }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_field_create_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload = parse_payload::<FeishuBitableFieldCreatePayload>(
-        "feishu.bitable.field.create",
-        request.payload,
-    )?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty(
-        "feishu.bitable.field.create",
-        "app_token",
-        &payload.app_token,
-    )?;
-    let table_id = require_non_empty("feishu.bitable.field.create", "table_id", &payload.table_id)?;
-    let field_name = require_non_empty(
-        "feishu.bitable.field.create",
-        "field_name",
-        &payload.field_name,
-    )?;
-    let field_type =
-        require_positive_i64("feishu.bitable.field.create", "type", payload.field_type)?;
-    let property = payload.property;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["bitable:app"], tool_name.as_str())?;
-        let field = bitable::create_bitable_field(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            &table_id,
-            &field_name,
-            field_type,
-            property,
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({ "field": field }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_field_list_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload = parse_payload::<FeishuBitableFieldListPayload>(
-        "feishu.bitable.field.list",
-        request.payload,
-    )?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token =
-        require_non_empty("feishu.bitable.field.list", "app_token", &payload.app_token)?;
-    let table_id = require_non_empty("feishu.bitable.field.list", "table_id", &payload.table_id)?;
-    let query = bitable::BitableFieldListQuery {
-        view_id: payload.view_id,
-        page_size: payload.page_size,
-        page_token: payload.page_token,
-    };
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["bitable:app"], tool_name.as_str())?;
-        let result = bitable::list_bitable_fields(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            &table_id,
-            &query,
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({
-                "fields": result.items,
-                "page_token": result.page_token,
-                "has_more": result.has_more,
-                "total": result.total,
-            }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_field_update_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload = parse_payload::<FeishuBitableFieldUpdatePayload>(
-        "feishu.bitable.field.update",
-        request.payload,
-    )?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty(
-        "feishu.bitable.field.update",
-        "app_token",
-        &payload.app_token,
-    )?;
-    let table_id = require_non_empty("feishu.bitable.field.update", "table_id", &payload.table_id)?;
-    let field_id = require_non_empty("feishu.bitable.field.update", "field_id", &payload.field_id)?;
-    let field_name = require_non_empty(
-        "feishu.bitable.field.update",
-        "field_name",
-        &payload.field_name,
-    )?;
-    let field_type =
-        require_positive_i64("feishu.bitable.field.update", "type", payload.field_type)?;
-    let property = payload.property;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["bitable:app"], tool_name.as_str())?;
-        let field = bitable::update_bitable_field(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            &table_id,
-            &field_id,
-            &field_name,
-            field_type,
-            property,
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({ "field": field }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_field_delete_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload = parse_payload::<FeishuBitableFieldDeletePayload>(
-        "feishu.bitable.field.delete",
-        request.payload,
-    )?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty(
-        "feishu.bitable.field.delete",
-        "app_token",
-        &payload.app_token,
-    )?;
-    let table_id = require_non_empty("feishu.bitable.field.delete", "table_id", &payload.table_id)?;
-    let field_id = require_non_empty("feishu.bitable.field.delete", "field_id", &payload.field_id)?;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["bitable:app"], tool_name.as_str())?;
-        let result = bitable::delete_bitable_field(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            &table_id,
-            &field_id,
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({
-                "deleted": result.deleted,
-                "field_id": result.field_id,
-            }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_view_create_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload = parse_payload::<FeishuBitableViewCreatePayload>(
-        "feishu.bitable.view.create",
-        request.payload,
-    )?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty(
-        "feishu.bitable.view.create",
-        "app_token",
-        &payload.app_token,
-    )?;
-    let table_id = require_non_empty("feishu.bitable.view.create", "table_id", &payload.table_id)?;
-    let view_name = require_non_empty(
-        "feishu.bitable.view.create",
-        "view_name",
-        &payload.view_name,
-    )?;
-    let view_type = payload.view_type;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["bitable:app"], tool_name.as_str())?;
-        let view = bitable::create_bitable_view(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            &table_id,
-            &view_name,
-            view_type.as_deref(),
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({ "view": view }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_view_get_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload =
-        parse_payload::<FeishuBitableViewGetPayload>("feishu.bitable.view.get", request.payload)?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty("feishu.bitable.view.get", "app_token", &payload.app_token)?;
-    let table_id = require_non_empty("feishu.bitable.view.get", "table_id", &payload.table_id)?;
-    let view_id = require_non_empty("feishu.bitable.view.get", "view_id", &payload.view_id)?;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["bitable:app"], tool_name.as_str())?;
-        let view = bitable::get_bitable_view(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            &table_id,
-            &view_id,
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({ "view": view }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_view_list_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload =
-        parse_payload::<FeishuBitableViewListPayload>("feishu.bitable.view.list", request.payload)?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token = require_non_empty("feishu.bitable.view.list", "app_token", &payload.app_token)?;
-    let table_id = require_non_empty("feishu.bitable.view.list", "table_id", &payload.table_id)?;
-    let query = bitable::BitableViewListQuery {
-        page_size: payload.page_size,
-        page_token: payload.page_token,
-    };
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["bitable:app"], tool_name.as_str())?;
-        let result = bitable::list_bitable_views(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            &table_id,
-            &query,
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({
-                "views": result.items,
-                "page_token": result.page_token,
-                "has_more": result.has_more,
-                "total": result.total,
-            }),
-        ))
-    })
-}
-
-fn execute_feishu_bitable_view_patch_tool_with_config(
-    request: ToolCoreRequest,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> Result<ToolCoreOutcome, String> {
-    let payload = parse_payload::<FeishuBitableViewPatchPayload>(
-        "feishu.bitable.view.patch",
-        request.payload,
-    )?;
-    let context = load_feishu_tool_context(
-        config,
-        requested_account_id(payload.selector.account_id.as_deref(), &payload.internal),
-    )?;
-    let grant = require_selected_grant(&context, payload.selector.open_id.as_deref())?;
-    let app_token =
-        require_non_empty("feishu.bitable.view.patch", "app_token", &payload.app_token)?;
-    let table_id = require_non_empty("feishu.bitable.view.patch", "table_id", &payload.table_id)?;
-    let view_id = require_non_empty("feishu.bitable.view.patch", "view_id", &payload.view_id)?;
-    let view_name =
-        require_non_empty("feishu.bitable.view.patch", "view_name", &payload.view_name)?;
-    let tool_name = request.tool_name;
-
-    run_feishu_future(async move {
-        let grant = crate::channel::feishu::api::ensure_fresh_user_grant(
-            &context.client,
-            &context.store,
-            &grant,
-        )
-        .await?;
-        ensure_any_required_scope(&grant, &["bitable:app"], tool_name.as_str())?;
-        let view = bitable::patch_bitable_view(
-            &context.client,
-            &grant.access_token,
-            &app_token,
-            &table_id,
-            &view_id,
-            &view_name,
-        )
-        .await?;
-
-        Ok(ok_outcome(
-            tool_name.as_str(),
-            context.configured_account_label.as_str(),
-            context.account_id.as_str(),
-            &grant.principal,
-            json!({ "view": view }),
         ))
     })
 }
@@ -5065,917 +2288,5 @@ fn execute_feishu_calendar_freebusy_tool_with_config(
     })
 }
 
-fn parse_payload<T>(tool_name: &str, payload: serde_json::Value) -> Result<T, String>
-where
-    T: DeserializeOwned,
-{
-    serde_json::from_value(payload)
-        .map_err(|error| format!("{tool_name} payload validation failed: {error}"))
-}
-
-fn load_feishu_tool_context(
-    config: &super::runtime_config::ToolRuntimeConfig,
-    requested_account_id: Option<&str>,
-) -> CliResult<FeishuToolContext> {
-    let Some(runtime) = config.feishu.as_ref() else {
-        return Err(
-            "feishu tool runtime is unavailable; configure feishu credentials and integration storage first"
-                .to_owned(),
-        );
-    };
-    let resolved = crate::channel::feishu::api::resolve_requested_feishu_account(
-        &runtime.channel,
-        trimmed_opt(requested_account_id),
-        "set payload.account_id to one of those configured accounts to disambiguate the Feishu tool request",
-    )?;
-    let client = FeishuClient::from_configs(&resolved, &runtime.integration)?;
-    let configured_account_id = resolved.configured_account_id.clone();
-    let configured_account_label = resolved.configured_account_label.clone();
-    let account_id = resolved.account.id.clone();
-    let receive_id_type = resolved.receive_id_type;
-    let store = FeishuTokenStore::new(runtime.integration.resolved_sqlite_path());
-
-    Ok(FeishuToolContext {
-        configured_account_id,
-        configured_account_label,
-        account_id,
-        receive_id_type,
-        client,
-        store,
-    })
-}
-
-fn require_selected_grant(
-    context: &FeishuToolContext,
-    open_id: Option<&str>,
-) -> CliResult<FeishuGrant> {
-    let resolution = crate::channel::feishu::api::resolve_grant_selection(
-        &context.store,
-        context.account_id.as_str(),
-        trimmed_opt(open_id),
-    )?;
-    if let Some(grant) = resolution.selected_grant().cloned() {
-        return Ok(grant);
-    }
-    Err(
-        crate::channel::feishu::api::describe_grant_selection_error_for_display(
-            context.account_id.as_str(),
-            context.configured_account_id.as_str(),
-            &resolution,
-        )
-        .unwrap_or_else(|| {
-            format!(
-                "no stored Feishu grant for account `{}`; run `{} feishu auth start --account {}` first",
-                context.configured_account_id,
-                crate::config::active_cli_command_name(),
-                context.configured_account_id
-            )
-        }),
-    )
-}
-
-fn require_non_empty(tool_name: &str, field: &str, value: &str) -> CliResult<String> {
-    let trimmed = value.trim();
-    if trimmed.is_empty() {
-        return Err(format!("{tool_name} requires payload.{field}"));
-    }
-    Ok(trimmed.to_owned())
-}
-
-fn require_positive_i64(tool_name: &str, field: &str, value: i64) -> CliResult<i64> {
-    if value > 0 {
-        return Ok(value);
-    }
-
-    Err(format!(
-        "{tool_name} invalid payload.{field}: expected positive integer, got {value}"
-    ))
-}
-
-fn resolve_feishu_doc_content_type(
-    tool_name: &str,
-    has_content: bool,
-    raw: Option<&str>,
-) -> CliResult<Option<&'static str>> {
-    match trimmed_opt(raw) {
-        Some(value) => match value.to_ascii_lowercase().as_str() {
-            "markdown" => Ok(Some("markdown")),
-            "html" => Ok(Some("html")),
-            other => Err(format!(
-                "unsupported feishu document content_type `{other}`; expected `markdown` or `html`"
-            )),
-        },
-        None if !has_content && raw.is_some() => Err(format!(
-            "{tool_name} payload.content_type requires payload.content or payload.content_path"
-        )),
-        None => Ok(None),
-    }
-}
-
-fn prepare_feishu_doc_tool_content(
-    tool_name: &str,
-    content: Option<&str>,
-    content_path: Option<&str>,
-    content_type: Option<&str>,
-    required: bool,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> CliResult<Option<PreparedFeishuDocContent>> {
-    let inline_content = trimmed_opt(content).map(ToOwned::to_owned);
-    let file_path = trimmed_opt(content_path);
-    if inline_content.is_some() && file_path.is_some() {
-        return Err(format!(
-            "{tool_name} accepts either payload.content or payload.content_path, not both"
-        ));
-    }
-
-    let has_content = inline_content.is_some() || file_path.is_some();
-    let explicit_content_type =
-        resolve_feishu_doc_content_type(tool_name, has_content, content_type)?;
-
-    match (inline_content, file_path) {
-        (Some(content), None) => Ok(Some(PreparedFeishuDocContent {
-            content,
-            content_type: explicit_content_type.unwrap_or("markdown"),
-        })),
-        (None, Some(path)) => {
-            let content =
-                read_safe_tool_text_file(tool_name, "payload.content_path", path, config)?;
-            Ok(Some(PreparedFeishuDocContent {
-                content,
-                content_type: explicit_content_type
-                    .unwrap_or_else(|| infer_feishu_doc_content_type_from_path(Path::new(path))),
-            }))
-        }
-        (None, None) if required => Err(format!(
-            "{tool_name} requires payload.content or payload.content_path"
-        )),
-        (None, None) => Ok(None),
-        (Some(_), Some(_)) => Err(format!(
-            "{tool_name} accepts either payload.content or payload.content_path, not both"
-        )),
-    }
-}
-
-fn infer_feishu_doc_content_type_from_path(path: &Path) -> &'static str {
-    match path
-        .extension()
-        .and_then(|value| value.to_str())
-        .map(|value| value.to_ascii_lowercase())
-        .as_deref()
-    {
-        Some("html") | Some("htm") => "html",
-        Some("md") | Some("markdown") => "markdown",
-        _ => "markdown",
-    }
-}
-
-fn require_non_empty_with_fallback(
-    tool_name: &str,
-    field: &str,
-    value: Option<&str>,
-    fallback: Option<&str>,
-) -> CliResult<String> {
-    value
-        .and_then(|value| trimmed_opt(Some(value)))
-        .or_else(|| fallback.and_then(|value| trimmed_opt(Some(value))))
-        .map(str::to_owned)
-        .ok_or_else(|| format!("{tool_name} requires payload.{field}"))
-}
-
-fn normalize_open_ids<'a, I>(values: I) -> Vec<String>
-where
-    I: IntoIterator<Item = &'a str>,
-{
-    let mut seen = std::collections::BTreeSet::new();
-    let mut normalized = Vec::new();
-    for value in values {
-        let trimmed = value.trim();
-        if trimmed.is_empty() || !seen.insert(trimmed.to_owned()) {
-            continue;
-        }
-        normalized.push(trimmed.to_owned());
-    }
-    normalized
-}
-
-fn requested_account_id<'a>(
-    explicit: Option<&'a str>,
-    internal: &'a LoongInternalToolPayload,
-) -> Option<&'a str> {
-    explicit.or_else(|| internal.ingress_requested_account_id())
-}
-
-fn resolve_message_resource_selection(
-    tool_name: &str,
-    effective_message_id: &str,
-    payload_file_key: &str,
-    payload_resource_type: &str,
-    internal: &LoongInternalToolPayload,
-) -> CliResult<(String, String)> {
-    let explicit_file_key = trimmed_opt(Some(payload_file_key));
-    let explicit_resource_type = trimmed_opt(Some(payload_resource_type))
-        .map(normalize_message_resource_type_alias)
-        .transpose()
-        .map_err(|error| format!("{tool_name} invalid payload.type: {error}"))?;
-    let ingress_message_override = ingress_message_override_reason(internal, effective_message_id);
-    let ingress_resources = ingress_resources_for_effective_message(internal, effective_message_id);
-    let ingress_resource = match (explicit_file_key, explicit_resource_type.as_deref()) {
-        (None, None) => {
-            single_ingress_resource_for_selection(tool_name, ingress_resources.as_slice())?
-        }
-        (Some(explicit_file_key), None) => infer_ingress_resource_from_file_key(
-            tool_name,
-            explicit_file_key,
-            ingress_resources.as_slice(),
-        )?,
-        (None, Some(explicit_resource_type)) => infer_ingress_resource_from_type(
-            tool_name,
-            explicit_resource_type,
-            ingress_resources.as_slice(),
-        )?,
-        (Some(explicit_file_key), Some(explicit_resource_type)) => {
-            validate_explicit_ingress_resource_pair(
-                tool_name,
-                explicit_file_key,
-                explicit_resource_type,
-                ingress_resources.as_slice(),
-            )?;
-            None
-        }
-    };
-
-    if let Some(current_ingress_message_id) = ingress_message_override {
-        match (explicit_file_key, explicit_resource_type.as_deref()) {
-            (None, None) => {
-                return Err(format!(
-                    "{tool_name} requires payload.file_key and payload.type because payload.message_id `{effective_message_id}` differs from current Feishu ingress message `{current_ingress_message_id}`; current ingress resource defaults only apply when payload.message_id is omitted or matches the current message"
-                ));
-            }
-            (None, Some(_)) => {
-                return Err(format!(
-                    "{tool_name} requires payload.file_key because payload.message_id `{effective_message_id}` differs from current Feishu ingress message `{current_ingress_message_id}`; current ingress resource defaults only apply when payload.message_id is omitted or matches the current message"
-                ));
-            }
-            (Some(_), None) => {
-                return Err(format!(
-                    "{tool_name} requires payload.type because payload.message_id `{effective_message_id}` differs from current Feishu ingress message `{current_ingress_message_id}`; current ingress resource defaults only apply when payload.message_id is omitted or matches the current message"
-                ));
-            }
-            (Some(_), Some(_)) => {}
-        }
-    }
-
-    let file_key = require_non_empty_with_fallback(
-        tool_name,
-        "file_key",
-        explicit_file_key,
-        ingress_resource
-            .as_ref()
-            .map(|resource| resource.file_key.as_str()),
-    )?;
-    let resource_type = require_non_empty_with_fallback(
-        tool_name,
-        "type",
-        explicit_resource_type.as_deref(),
-        ingress_resource
-            .as_ref()
-            .map(|resource| resource.resource_type.as_str()),
-    )?;
-    Ok((file_key, resource_type))
-}
-
-fn ingress_resources_for_effective_message(
-    internal: &LoongInternalToolPayload,
-    effective_message_id: &str,
-) -> Vec<FeishuInternalIngressResolvedResource> {
-    if internal
-        .ingress_message_id()
-        .is_some_and(|message_id| message_id == effective_message_id)
-    {
-        return internal.ingress_resources();
-    }
-    Vec::new()
-}
-
-fn ingress_message_override_reason<'a>(
-    internal: &'a LoongInternalToolPayload,
-    effective_message_id: &str,
-) -> Option<&'a str> {
-    internal
-        .ingress_message_id()
-        .filter(|message_id| *message_id != effective_message_id)
-}
-
-fn single_ingress_resource_for_selection(
-    tool_name: &str,
-    ingress_resources: &[FeishuInternalIngressResolvedResource],
-) -> CliResult<Option<FeishuInternalIngressResolvedResource>> {
-    match ingress_resources {
-        [] => Ok(None),
-        [resource] => Ok(Some(resource.clone())),
-        _ => Err(format!(
-            "{tool_name} requires payload.file_key and payload.type when current Feishu ingress carries multiple Feishu message resources; available ingress resources: {}. If the current Feishu ingress summary includes resource_inventory, choose one entry and copy its file_key plus payload_type.",
-            describe_ingress_resources(ingress_resources)
-        )),
-    }
-}
-
-fn validate_explicit_ingress_resource_pair(
-    tool_name: &str,
-    explicit_file_key: &str,
-    explicit_resource_type: &str,
-    ingress_resources: &[FeishuInternalIngressResolvedResource],
-) -> CliResult<()> {
-    if ingress_resources.is_empty() {
-        return Ok(());
-    }
-
-    if ingress_resources.iter().any(|resource| {
-        resource.file_key == explicit_file_key
-            && normalize_ingress_message_resource_type(tool_name, resource)
-                .is_ok_and(|resource_type| resource_type == explicit_resource_type)
-    }) {
-        return Ok(());
-    }
-
-    let matching_file_key = ingress_resources
-        .iter()
-        .find(|resource| resource.file_key == explicit_file_key);
-    if let Some(resource) = matching_file_key {
-        return Err(format!(
-            "{tool_name} payload.type conflicts with the current Feishu ingress resource selected by payload.file_key ({}); choose one entry from resource_inventory and copy its payload_type, or override both payload.message_id and payload.file_key when targeting a different Feishu message resource",
-            describe_ingress_resource(resource)
-        ));
-    }
-
-    let matching_type = ingress_resources
-        .iter()
-        .filter(|resource| {
-            normalize_ingress_message_resource_type(tool_name, resource)
-                .is_ok_and(|resource_type| resource_type == explicit_resource_type)
-        })
-        .collect::<Vec<_>>();
-    if !matching_type.is_empty() {
-        return Err(format!(
-            "{tool_name} payload.file_key `{explicit_file_key}` does not match the current Feishu ingress resource(s) selected by payload.type: {}. Choose one entry from resource_inventory and copy its file_key, or override both payload.message_id and payload.file_key when targeting a different Feishu message resource",
-            describe_ingress_resource_matches(matching_type.as_slice())
-        ));
-    }
-
-    Err(format!(
-        "{tool_name} payload.file_key `{explicit_file_key}` and payload.type `{explicit_resource_type}` did not match any current Feishu ingress resource; available ingress resources: {}. Choose one entry from resource_inventory and copy its file_key plus payload_type, or override payload.message_id when targeting a different Feishu message resource",
-        describe_ingress_resources(ingress_resources)
-    ))
-}
-
-fn infer_ingress_resource_from_file_key(
-    tool_name: &str,
-    explicit_file_key: &str,
-    ingress_resources: &[FeishuInternalIngressResolvedResource],
-) -> CliResult<Option<FeishuInternalIngressResolvedResource>> {
-    match ingress_resources {
-        [] => Ok(None),
-        [resource] => {
-            if explicit_file_key != resource.file_key {
-                return Err(format!(
-                    "{tool_name} payload.file_key conflicts with the current Feishu ingress resource ({}); provide payload.type explicitly to override ingress defaults or omit payload.file_key to use ingress defaults",
-                    describe_ingress_resource(resource)
-                ));
-            }
-            Ok(Some(resource.clone()))
-        }
-        _ => {
-            let matches = ingress_resources
-                .iter()
-                .filter(|resource| resource.file_key == explicit_file_key)
-                .collect::<Vec<_>>();
-            match matches.as_slice() {
-                [] => Err(format!(
-                    "{tool_name} payload.file_key `{explicit_file_key}` did not match any current Feishu ingress resource; available ingress resources: {}. Provide payload.type explicitly to override ingress defaults or choose one entry from resource_inventory and copy its file_key plus payload_type.",
-                    describe_ingress_resources(ingress_resources)
-                )),
-                [resource] => Ok(Some((*resource).clone())),
-                _ => Err(format!(
-                    "{tool_name} payload.file_key matches multiple current Feishu ingress resources: {}. Provide payload.type explicitly to disambiguate or choose one entry from resource_inventory and copy its payload_type.",
-                    describe_ingress_resource_matches(matches.as_slice())
-                )),
-            }
-        }
-    }
-}
-
-fn infer_ingress_resource_from_type(
-    tool_name: &str,
-    explicit_resource_type: &str,
-    ingress_resources: &[FeishuInternalIngressResolvedResource],
-) -> CliResult<Option<FeishuInternalIngressResolvedResource>> {
-    match ingress_resources {
-        [] => Ok(None),
-        [resource] => {
-            let ingress_resource_type =
-                normalize_ingress_message_resource_type(tool_name, resource)?;
-            if explicit_resource_type != ingress_resource_type {
-                return Err(format!(
-                    "{tool_name} payload.type conflicts with the current Feishu ingress resource ({}); provide payload.file_key explicitly to override ingress defaults or omit payload.type to use ingress defaults",
-                    describe_ingress_resource(resource)
-                ));
-            }
-            Ok(Some(resource.clone()))
-        }
-        _ => {
-            let mut matches = Vec::new();
-            for resource in ingress_resources {
-                if explicit_resource_type
-                    == normalize_ingress_message_resource_type(tool_name, resource)?
-                {
-                    matches.push(resource);
-                }
-            }
-            match matches.as_slice() {
-                [] => Err(format!(
-                    "{tool_name} payload.type `{explicit_resource_type}` did not match any current Feishu ingress resource; available ingress resources: {}. Provide payload.file_key explicitly to override ingress defaults or choose one entry from resource_inventory and copy its file_key plus payload_type.",
-                    describe_ingress_resources(ingress_resources)
-                )),
-                [resource] => Ok(Some((*resource).clone())),
-                _ => Err(format!(
-                    "{tool_name} payload.type matches multiple current Feishu ingress resources: {}. Provide payload.file_key explicitly to disambiguate and choose one entry from resource_inventory.",
-                    describe_ingress_resource_matches(matches.as_slice())
-                )),
-            }
-        }
-    }
-}
-
-fn normalize_ingress_message_resource_type(
-    tool_name: &str,
-    resource: &FeishuInternalIngressResolvedResource,
-) -> CliResult<String> {
-    normalize_message_resource_type_alias(resource.resource_type.as_str())
-        .map_err(|error| format!("{tool_name} invalid ingress resource type: {error}"))
-}
-
-fn normalize_message_resource_type_alias(value: &str) -> CliResult<String> {
-    value
-        .parse::<FeishuMessageResourceType>()
-        .map(|resource_type| resource_type.as_api_value().to_owned())
-}
-
-fn prepare_feishu_tool_media(
-    tool_name: &str,
-    image_key: Option<&str>,
-    image_path: Option<&str>,
-    file_key: Option<&str>,
-    file_path: Option<&str>,
-    file_type: Option<&str>,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> CliResult<PreparedFeishuToolMedia> {
-    ensure_tool_media_source_exclusive(
-        tool_name,
-        "payload.image_key",
-        image_key,
-        "payload.image_path",
-        image_path,
-    )?;
-    ensure_tool_media_source_exclusive(
-        tool_name,
-        "payload.file_key",
-        file_key,
-        "payload.file_path",
-        file_path,
-    )?;
-    if trimmed_opt(file_type).is_some() && trimmed_opt(file_path).is_none() {
-        return Err(format!(
-            "{tool_name} only allows payload.file_type with payload.file_path"
-        ));
-    }
-
-    let image_key = trimmed_opt(image_key).map(ToOwned::to_owned);
-    let file_key = trimmed_opt(file_key).map(ToOwned::to_owned);
-    let image_upload = match trimmed_opt(image_path) {
-        Some(path) => Some(read_safe_tool_media_file(
-            tool_name,
-            "payload.image_path",
-            path,
-            config,
-        )?),
-        None => None,
-    };
-    let file_upload = match trimmed_opt(file_path) {
-        Some(path) => {
-            let upload = read_safe_tool_media_file(tool_name, "payload.file_path", path, config)?;
-            Some(PreparedFeishuToolFileUpload {
-                file_name: upload.file_name,
-                bytes: upload.bytes,
-                file_type: trimmed_opt(file_type)
-                    .unwrap_or(media::FEISHU_DEFAULT_MESSAGE_FILE_TYPE)
-                    .to_owned(),
-            })
-        }
-        None => None,
-    };
-
-    Ok(PreparedFeishuToolMedia {
-        image_key,
-        image_upload,
-        file_key,
-        file_upload,
-    })
-}
-
-fn validate_feishu_tool_message_body_fields(
-    tool_name: &str,
-    text: Option<&str>,
-    as_card: bool,
-    post: Option<&Value>,
-    image_key: Option<&str>,
-    image_path: Option<&str>,
-    file_key: Option<&str>,
-    file_path: Option<&str>,
-) -> CliResult<()> {
-    messages::resolve_outbound_message_body(
-        tool_name,
-        "payload.text",
-        "payload.as_card",
-        "payload.post",
-        "payload.image_key/payload.image_path",
-        "payload.file_key/payload.file_path",
-        text,
-        as_card,
-        post,
-        trimmed_opt(image_key).or_else(|| trimmed_opt(image_path).map(|_| "__image_path__")),
-        trimmed_opt(file_key).or_else(|| trimmed_opt(file_path).map(|_| "__file_path__")),
-    )
-    .map(|_| ())
-}
-
-async fn resolve_prepared_feishu_tool_media(
-    client: &FeishuClient,
-    tenant_access_token: &str,
-    prepared: PreparedFeishuToolMedia,
-) -> CliResult<ResolvedFeishuToolMedia> {
-    let image_key = match (prepared.image_key, prepared.image_upload) {
-        (Some(image_key), None) => Some(image_key),
-        (None, Some(upload)) => Some(
-            media::upload_message_image(
-                client,
-                tenant_access_token,
-                upload.file_name.as_str(),
-                upload.bytes,
-            )
-            .await?
-            .image_key,
-        ),
-        (Some(_), Some(_)) => {
-            return Err(
-                "feishu tool media preparation allowed both image_key and image_upload".to_owned(),
-            );
-        }
-        (None, None) => None,
-    };
-    let file_key = match (prepared.file_key, prepared.file_upload) {
-        (Some(file_key), None) => Some(file_key),
-        (None, Some(upload)) => Some(
-            media::upload_message_file(
-                client,
-                tenant_access_token,
-                upload.file_name.as_str(),
-                upload.bytes,
-                upload.file_type.as_str(),
-                None,
-            )
-            .await?
-            .file_key,
-        ),
-        (Some(_), Some(_)) => {
-            return Err(
-                "feishu tool media preparation allowed both file_key and file_upload".to_owned(),
-            );
-        }
-        (None, None) => None,
-    };
-
-    Ok(ResolvedFeishuToolMedia {
-        image_key,
-        file_key,
-    })
-}
-
-fn ensure_tool_media_source_exclusive(
-    tool_name: &str,
-    key_field: &str,
-    key: Option<&str>,
-    path_field: &str,
-    path: Option<&str>,
-) -> CliResult<()> {
-    if trimmed_opt(key).is_some() && trimmed_opt(path).is_some() {
-        return Err(format!(
-            "{tool_name} accepts either {key_field} or {path_field}, not both"
-        ));
-    }
-    Ok(())
-}
-
-#[cfg(feature = "tool-file")]
-fn read_safe_tool_text_file(
-    tool_name: &str,
-    field: &str,
-    raw_path: &str,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> CliResult<String> {
-    let resolved = super::file::resolve_safe_file_path_with_config(raw_path, config)?;
-    let bytes = fs::read(&resolved).map_err(|error| {
-        format!(
-            "{tool_name} failed to read {} `{}`: {error}",
-            field,
-            resolved.display()
-        )
-    })?;
-    if bytes.is_empty() {
-        return Err(format!(
-            "{tool_name} requires {} `{}` to be non-empty UTF-8 text",
-            field,
-            resolved.display()
-        ));
-    }
-    let content = String::from_utf8(bytes).map_err(|error| {
-        format!(
-            "{tool_name} requires {} `{}` to contain valid UTF-8 text: {error}",
-            field,
-            resolved.display()
-        )
-    })?;
-    let trimmed = content.trim();
-    if trimmed.is_empty() {
-        return Err(format!(
-            "{tool_name} requires {} `{}` to be non-empty UTF-8 text",
-            field,
-            resolved.display()
-        ));
-    }
-    Ok(trimmed.to_owned())
-}
-
-#[cfg(not(feature = "tool-file"))]
-fn read_safe_tool_text_file(
-    tool_name: &str,
-    field: &str,
-    raw_path: &str,
-    _config: &super::runtime_config::ToolRuntimeConfig,
-) -> CliResult<String> {
-    let _ = raw_path;
-    Err(format!(
-        "{tool_name} does not support {field} unless feature `tool-file` is enabled"
-    ))
-}
-
-#[cfg(feature = "tool-file")]
-fn read_safe_tool_media_file(
-    tool_name: &str,
-    field: &str,
-    raw_path: &str,
-    config: &super::runtime_config::ToolRuntimeConfig,
-) -> CliResult<PreparedFeishuToolUpload> {
-    let resolved = super::file::resolve_safe_file_path_with_config(raw_path, config)?;
-    let file_name = resolved
-        .file_name()
-        .and_then(|value| value.to_str())
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(ToOwned::to_owned)
-        .ok_or_else(|| format!("{tool_name} requires {field} to include a file name"))?;
-    let bytes = fs::read(&resolved).map_err(|error| {
-        format!(
-            "{tool_name} failed to read {} `{}`: {error}",
-            field,
-            resolved.display()
-        )
-    })?;
-    if bytes.is_empty() {
-        return Err(format!(
-            "{tool_name} requires {} `{}` to be non-empty",
-            field,
-            resolved.display()
-        ));
-    }
-    Ok(PreparedFeishuToolUpload { file_name, bytes })
-}
-
-#[cfg(not(feature = "tool-file"))]
-fn read_safe_tool_media_file(
-    tool_name: &str,
-    field: &str,
-    raw_path: &str,
-    _config: &super::runtime_config::ToolRuntimeConfig,
-) -> CliResult<PreparedFeishuToolUpload> {
-    let _ = raw_path;
-    Err(format!(
-        "{tool_name} does not support {field} unless feature `tool-file` is enabled"
-    ))
-}
-
-fn search_chat_scope(payload: &FeishuMessagesSearchPayload) -> Vec<String> {
-    let explicit = payload
-        .chat_ids
-        .iter()
-        .map(String::as_str)
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-        .map(ToOwned::to_owned)
-        .collect::<Vec<_>>();
-    if !explicit.is_empty() {
-        return explicit;
-    }
-
-    payload
-        .internal
-        .ingress_conversation_id()
-        .map(|conversation_id| vec![conversation_id.to_owned()])
-        .unwrap_or_default()
-}
-
 #[cfg(test)]
-fn push_feishu_registry_entry(
-    entries: &mut Vec<super::ToolRegistryEntry>,
-    name: &'static str,
-    description: &'static str,
-) {
-    let entry = super::ToolRegistryEntry {
-        name: name.to_owned(),
-        description: description.to_owned(),
-    };
-    entries.push(entry);
-}
-
-fn push_feishu_provider_tool_definition(
-    tools: &mut Vec<Value>,
-    name: &'static str,
-    description: &'static str,
-    parameters: Value,
-) {
-    tools.push(json!({
-        "type": "function",
-        "function": {
-            "name": name,
-            "description": description,
-            "parameters": parameters,
-        }
-    }));
-}
-
-fn feishu_provider_tool_function_name(tool: &Value) -> &str {
-    tool.get("function")
-        .and_then(|value| value.get("name"))
-        .and_then(Value::as_str)
-        .unwrap_or("")
-}
-
-fn ensure_required_scopes(
-    grant: &FeishuGrant,
-    required: &[&str],
-    tool_name: &str,
-) -> CliResult<()> {
-    let missing = required
-        .iter()
-        .copied()
-        .filter(|scope| !grant.scopes.contains(scope))
-        .collect::<Vec<_>>();
-    if missing.is_empty() {
-        return Ok(());
-    }
-
-    Err(format!(
-        "{tool_name} requires Feishu scopes [{}] for `{}`; update Feishu config if needed and rerun `loong feishu auth start --account <account>`",
-        missing.join(", "),
-        grant.principal.storage_key()
-    ))
-}
-
-fn ensure_any_required_scope(
-    grant: &FeishuGrant,
-    accepted: &[&str],
-    tool_name: &str,
-) -> CliResult<()> {
-    if accepted
-        .iter()
-        .copied()
-        .any(|scope| grant.scopes.contains(scope))
-    {
-        return Ok(());
-    }
-
-    Err(format!(
-        "{tool_name} requires at least one Feishu scope [{}] for `{}`; update Feishu config if needed and rerun `loong feishu auth start --account <account>`",
-        accepted.join(", "),
-        grant.principal.storage_key()
-    ))
-}
-
-fn ok_outcome(
-    tool_name: &str,
-    configured_account: &str,
-    account_id: &str,
-    principal: &FeishuUserPrincipal,
-    payload: serde_json::Value,
-) -> ToolCoreOutcome {
-    let mut body = json!({
-        "adapter": "core-tools",
-        "tool_name": tool_name,
-        "configured_account": configured_account,
-        "account_id": account_id,
-        "principal": principal,
-    });
-    if let Some(object) = body.as_object_mut()
-        && let Some(extra) = payload.as_object()
-    {
-        for (key, value) in extra {
-            object.insert(key.clone(), value.clone());
-        }
-    }
-    ToolCoreOutcome {
-        status: "ok".to_owned(),
-        payload: body,
-    }
-}
-
-fn ok_outcome_without_principal(
-    tool_name: &str,
-    configured_account: &str,
-    account_id: &str,
-    payload: serde_json::Value,
-) -> ToolCoreOutcome {
-    let mut body = json!({
-        "adapter": "core-tools",
-        "tool_name": tool_name,
-        "configured_account": configured_account,
-        "account_id": account_id,
-    });
-    if let Some(object) = body.as_object_mut()
-        && let Some(extra) = payload.as_object()
-    {
-        for (key, value) in extra {
-            object.insert(key.clone(), value.clone());
-        }
-    }
-    ToolCoreOutcome {
-        status: "ok".to_owned(),
-        payload: body,
-    }
-}
-
-fn run_feishu_future<F>(future: F) -> CliResult<ToolCoreOutcome>
-where
-    F: Future<Output = CliResult<ToolCoreOutcome>> + Send + 'static,
-{
-    std::thread::spawn(move || {
-        let runtime = tokio::runtime::Builder::new_current_thread()
-            .enable_all()
-            .build()
-            .map_err(|error| format!("build feishu tool runtime failed: {error}"))?;
-        runtime.block_on(future)
-    })
-    .join()
-    .map_err(|error| format!("feishu tool execution thread panicked: {error:?}"))?
-}
-
-fn trimmed_opt(value: Option<&str>) -> Option<&str> {
-    value.map(str::trim).filter(|value| !value.is_empty())
-}
-
-#[cfg(test)]
-mod payload_tests {
-    use super::*;
-
-    #[test]
-    fn feishu_calendar_primary_get_payload_accepts_selector_and_user_id_type() {
-        let payload: FeishuCalendarPrimaryGetPayload = serde_json::from_value(json!({
-            "account_id": "acct-001",
-            "open_id": "ou_abc",
-            "user_id_type": "union_id"
-        }))
-        .expect("primary get payload parses");
-        assert_eq!(payload.selector.account_id.as_deref(), Some("acct-001"));
-        assert_eq!(payload.selector.open_id.as_deref(), Some("ou_abc"));
-        assert_eq!(payload.user_id_type.as_deref(), Some("union_id"));
-    }
-
-    #[test]
-    fn feishu_calendar_primary_get_payload_defaults_to_empty() {
-        let payload: FeishuCalendarPrimaryGetPayload =
-            serde_json::from_value(json!({})).expect("empty primary get payload parses");
-        assert!(payload.selector.account_id.is_none());
-        assert!(payload.selector.open_id.is_none());
-        assert!(payload.user_id_type.is_none());
-    }
-
-    #[test]
-    fn feishu_calendar_primary_get_payload_rejects_unknown_fields() {
-        let err = serde_json::from_value::<FeishuCalendarPrimaryGetPayload>(json!({
-            "unexpected_field": true
-        }))
-        .expect_err("unknown fields must be rejected");
-        assert!(
-            err.to_string().contains("unexpected_field"),
-            "error should mention the unknown field, got: {err}"
-        );
-    }
-}
+mod payload_tests;
